feat(cache): Redesign cache JSON around python modules rather than oneil models
