+++ src-rs/oneil_py_call_cache/src/file.rs
@@ -1,41 +1,41 @@
 //! On-disk cache file: [`FileCache`] and JSON load/save.
 
-use std::collections::BTreeMap;
-use std::fs::File;
-use std::path::Path;
+use std::collections::BTreeSet;
+use std::path::{Path, PathBuf};
+use std::{collections::BTreeMap, fs::File};
 
-use oneil_shared::paths::PythonPath;
-use oneil_shared::symbols::{ParameterName, TestIndex};
-use serde::{Deserialize, Serialize};
+use oneil_shared::{paths::PythonPath, symbols::PyFunctionName};
+use serde::{Deserialize, Deserializer, Serialize, Serializer, de};
 
-use crate::FunctionCall;
-use crate::error::{ReadCacheError, WriteCacheError};
-use crate::imports::ImportEntry;
+use crate::{FunctionCall, ReadCacheError, WriteCacheError};
 
 const ONEIL_VERSION: &str = env!("CARGO_PKG_VERSION");
 
-/// On-disk cache for one source file: imported modules, parameter calls, and test calls.
+/// On-disk cache for one python module.
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct FileCache {
     /// The version of Oneil that created the cache.
     pub oneil_version: String,
-    /// Imported Python modules and their invalidation metadata.
-    pub imports: BTreeMap<PythonPath, ImportEntry>,
-    /// Function calls originating from named parameters.
-    pub parameters: BTreeMap<ParameterName, Vec<FunctionCall>>,
-    /// Function calls originating from tests, keyed by test index.
-    pub tests: BTreeMap<TestIndex, Vec<FunctionCall>>,
+    /// The path of the python module that was cached.
+    pub module_path: PythonPath,
+    /// The hash of the python module and its dependencies.
+    pub hash: ImportHash,
+    /// The local dependencies included in the combined hash.
+    pub dependencies: BTreeSet<PathBuf>,
+    /// The function calls that are cached for this module.
+    pub function_calls: BTreeMap<PyFunctionName, Vec<FunctionCall>>,
 }
 
 impl FileCache {
     /// Creates a new empty file cache.
     #[must_use]
-    pub fn new() -> Self {
+    pub fn new(module_path: PythonPath, hash: ImportHash, dependencies: BTreeSet<PathBuf>) -> Self {
         Self {
             oneil_version: ONEIL_VERSION.to_owned(),
-            imports: BTreeMap::new(),
-            parameters: BTreeMap::new(),
-            tests: BTreeMap::new(),
+            module_path,
+            hash,
+            dependencies,
+            function_calls: BTreeMap::new(),
         }
     }
 
@@ -67,8 +67,58 @@ impl FileCache {
     }
 }
 
-impl Default for FileCache {
-    fn default() -> Self {
-        Self::new()
+/// Fingerprint for a python module's sources (stored as raw `u64`, serialized as hex).
+#[derive(Clone, Copy, Debug, Eq, PartialEq)]
+pub struct ImportHash(u64);
+
+impl Serialize for ImportHash {
+    /// Writes this hash as a 16-digit lowercase hexadecimal string (JSON string).
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        serializer.serialize_str(&format!("{:016x}", self.0))
+    }
+}
+
+impl<'de> Deserialize<'de> for ImportHash {
+    /// Parses a base-16 string into a hash (no `0x` prefix).
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let s = String::deserialize(deserializer)?;
+
+        if s.is_empty() {
+            return Err(de::Error::custom("empty hexadecimal string"));
+        }
+
+        u64::from_str_radix(&s, 16)
+            .map(ImportHash)
+            .map_err(de::Error::custom)
+    }
+}
+
+impl PartialEq<u64> for ImportHash {
+    fn eq(&self, other: &u64) -> bool {
+        self.0 == *other
+    }
+}
+
+impl PartialEq<ImportHash> for u64 {
+    fn eq(&self, other: &ImportHash) -> bool {
+        *self == other.0
+    }
+}
+
+impl From<u64> for ImportHash {
+    fn from(hash: u64) -> Self {
+        Self(hash)
+    }
+}
+
+impl From<ImportHash> for u64 {
+    fn from(hash: ImportHash) -> Self {
+        hash.0
     }
 }
