+++ src-rs/oneil_py_call_cache/src/lib.rs
@@ -4,9 +4,7 @@
 mod error;
 mod file;
 mod function_call;
-mod imports;
 
 pub use error::{ReadCacheError, WriteCacheError};
 pub use file::FileCache;
 pub use function_call::{FunctionCall, FunctionCallResult};
-pub use imports::{ImportEntry, ImportHash};
